Change Percentage to u16

+++ include/ghostty.h
@@ -358,7 +358,7 @@ typedef enum {
 
 // apprt.action.SplitDirection
 typedef struct {
-  float percent;
+  uint16_t percent;
   ghostty_action_split_direction_e direction;
 } ghostty_action_split_s;
 
@@ -699,9 +699,9 @@ void ghostty_surface_mouse_scroll(ghostty_surface_t,
 void ghostty_surface_mouse_pressure(ghostty_surface_t, uint32_t, double);
 void ghostty_surface_ime_point(ghostty_surface_t, double*, double*);
 void ghostty_surface_request_close(ghostty_surface_t);
-void ghostty_surface_split(ghostty_surface_t, 
+void ghostty_surface_split(ghostty_surface_t,
                            ghostty_action_split_direction_e,
-                           float);
+                           uint16_t);
 void ghostty_surface_split_focus(ghostty_surface_t,
                                  ghostty_action_goto_split_e);
 void ghostty_surface_split_resize(ghostty_surface_t,
